dbt-bigquery/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
dynamic = ["version"]
name = "dbt-bigquery"
description = "The BigQuery adapter plugin for dbt"
readme = "README.md"
keywords = ["dbt", "adapter", "adapters", "database", "elt", "dbt-core", "dbt Core", "dbt Cloud", "dbt Labs", "bigquery", "google"]
requires-python = ">=3.9.0"
authors = [{ name = "dbt Labs", email = "info@dbtlabs.com" }]
maintainers = [{ name = "dbt Labs", email = "info@dbtlabs.com" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "dbt-common>=1.10,<2.0",
    "dbt-adapters>=1.7,<2.0",
    # 3.20 introduced pyarrow>=3.0 under the `pandas` extra
    "google-cloud-bigquery[pandas]>=3.0,<4.0",
    "google-cloud-storage~=2.4",
    "google-cloud-dataproc~=5.0",
    # ----
    # Expect compatibility with all new versions of these packages, so lower bounds only.
    "google-api-core>=2.11.0",
    # add dbt-core to ensure backwards compatibility of installation, this is not a functional dependency
    "dbt-core>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/dbt-labs/dbt-bigquery"
Documentation = "https://docs.getdbt.com"
Repository = "https://github.com/dbt-labs/dbt-bigquery.git"
Issues = "https://github.com/dbt-labs/dbt-bigquery/issues"
Changelog = "https://github.com/dbt-labs/dbt-bigquery/blob/main/CHANGELOG.md"

[tool.mypy]
mypy_path = "third-party-stubs/"

[tool.pytest.ini_options]
testpaths = ["tests/functional", "tests/unit"]
env_files = ["test.env"]
addopts = "-v -n auto"
color = true
filterwarnings = [
    "ignore:.*'soft_unicode' has been renamed to 'soft_str'*:DeprecationWarning",
    "ignore:unclosed file .*:ResourceWarning",
]
